Page components now implement their own query logic; dependencies on web app reduced to only DynamoDB scan and query methods.
