+++ web/components/app.jsx
@@ -14,45 +14,6 @@ class App extends Component {
   componentWillMount() {
     const app = this
 
-    function getCategoriesFromDynamo(dynamodb) {
-      const params = {
-        TableName: 'demo-ProductCategory-1',
-        AttributesToGet: ['category'],
-      }
-
-      return new Promise((resolve, reject) => {
-        dynamodb.scan(params, (err, data) => {
-          if (err) { reject(err) }
-          resolve(data)
-        })
-      })
-    }
-
-    function getProductsByCategoryFromDynamo(dynamodb, category) {
-      const params = {
-        ProjectionExpression: 'brand, description, #na',
-        TableName: 'demo-ProductCatalog-1',
-        IndexName: 'Category',
-        KeyConditionExpression: '#ct = :cat',
-        ExpressionAttributeNames: {
-          '#ct': 'category',
-          '#na': 'name',
-        },
-        ExpressionAttributeValues: {
-          ':cat': {
-            S: category,
-          },
-        },
-      }
-
-      return new Promise((resolve, reject) => {
-        dynamodb.query(params, (err, data) => {
-          if (err) { reject(err) }
-          resolve(data)
-        })
-      })
-    }
-
     function authAmazonLogin() {
       return new Promise((resolve, reject) => {
         amazon.Login.setClientId('amzn1.application-oa2-client.d38cc5e49f9343b5b7a718f9d21fc93a')
@@ -113,42 +74,14 @@ class App extends Component {
             secretAccessKey: role.Credentials.SecretAccessKey,
             sessionToken: role.Credentials.SessionToken,
           }
+
           AWS.config.region = 'us-west-2'
 
           const dynamodb = new AWS.DynamoDB()
 
-          app.fetchCategories = () => { // eslint-disable-line arrow-body-style
-            return getCategoriesFromDynamo(dynamodb)
-              .then((data) => {
-                const categoriesList = []
-                data.Items.forEach((item) => {
-                  categoriesList.push({
-                    name: item.category.S,
-                  })
-                })
-                return categoriesList
-              }, (error) => { throw new Error(error) })
-          }
-
-          app.fetchProductsByCategory = (category) => { // eslint-disable-line arrow-body-style
-            return getProductsByCategoryFromDynamo(dynamodb, category)
-              .then((data) => {
-                const productList = []
-                data.Items.forEach((item) => {
-                  productList.push({
-                    brand: item.brand.S,
-                    description: item.description.S,
-                    name: item.name.S,
-                  })
-                })
-                return productList
-              }, (error) => { throw new Error(error) })
-          }
-
           app.setState({
             loggedIn: true,
-            fetchCategories: app.fetchCategories,
-            fetchProductsByCategory: app.fetchProductsByCategory,
+            dynamoDb: dynamodb,
           })
         })
     }
@@ -163,9 +96,10 @@ class App extends Component {
     if (!this.state.loggedIn) {
       children = (<div>Logging in...</div>)
     } else {
+      // Maps properties to child components dynamically, allowing those properties to be bound once available.
       children = React.Children.map(this.props.children, child => React.cloneElement(child, { // eslint-disable-line react/prop-types
-        fetchCategories: app.state.fetchCategories,
-        fetchProductsByCategory: app.state.fetchProductsByCategory,
+        dynamoDb: app.state.dynamoDb,
+        // fetchProductsByCategory: app.state.fetchProductsByCategory,
       }))
     }
