+++ web/components/category/category-page.jsx
@@ -1,13 +1,17 @@
 import React, { Component, PropTypes } from 'react' // eslint-disable-line import/no-extraneous-dependencies
 import CategoryList from './category-list'
 
+// TODO: Further decompose using HOC https://facebook.github.io/react/docs/higher-order-components.html
+
 class CategoryPage extends Component {
   static propTypes = {
-    fetchCategories: PropTypes.func,
+    dynamoDb: PropTypes.shape({
+      scan: PropTypes.func.isRequired,
+    }),
   }
 
   static defaultProps = {
-    fetchCategories: () => [],
+    dynamoDb: null,
   }
 
   constructor(props) {
@@ -15,17 +19,52 @@ class CategoryPage extends Component {
     this.state = {}
   }
 
-  componentWillMount() {
-    this.props.fetchCategories()
-      .then((categories) => {
-        console.log(categories)
-        this.setState({
-          categoryList: categories,
-        })
+  getCategoriesFromDynamoAsync() {
+    const params = {
+      TableName: 'demo-ProductCategory-1',
+      AttributesToGet: ['category'],
+    }
+
+    return new Promise((resolve, reject) => {
+      this.props.dynamoDb.scan(params, (err, data) => {
+        if (err) { reject(err) }
+        resolve(data)
       })
+    })
+  }
+
+  getCategoriesAsync() {
+    return this.getCategoriesFromDynamoAsync()
+      .then((data) => {
+        const categoriesList = []
+        data.Items.forEach((item) => {
+          categoriesList.push({
+            name: item.category.S,
+          })
+        })
+        return categoriesList
+      }, (error) => { throw new Error(error) })
   }
 
   render() {
+    // No data connection available yet.
+    if (!this.props.dynamoDb) {
+      return (<div>Connecting...</div>)
+    }
+
+    // No categories yet.
+    if (!this.state.categoryList) {
+      this.getCategoriesAsync()
+        .then((categories) => {
+          console.log(categories)
+          this.setState({
+            categoryList: categories,
+          })
+        })
+
+      return (<div>Loading Categories...</div>)
+    }
+
     return (
       <div>
         <h4><em>Categories</em></h4>
